requirements.txt
python-telegram-bot>=20.7,<22
google-genai>=0.2.0
Pillow>=10.3.0
python-dotenv>=1.0.1
pytest>=8.2.0
streamlit>=1.36.0

